fix(test): Skip failing storage options query test on Cloudberry
Temporarily skips the `GetTableStorage` integration test on Cloudberry.
The test fails because Cloudberry, unlike GPDB 7, does not persist
default storage options to `pg_class.reloptions` when an append-optimized
table is created with only `appendonly=true`.

This inconsistency causes the test to fail as it expects the default
storage option string but receives an empty one from the query.

A TODO comment has been added to re-enable the test once the underlying
issue in Cloudberry is resolved.

See: https://github.com/apache/cloudberry/issues/1260