fix(backup): Use standard LOCK TABLE syntax for Cloudberry
Temporarily modifies the table locking logic to use a standard
`LOCK TABLE ... IN ACCESS SHARE MODE` command when running against
Cloudberry.

This change is necessary because Cloudberry, which is based on
PostgreSQL 14, does not support the `COORDINATOR ONLY` clause that was
introduced in GPDB 7 as a performance optimization. The use of this
unsupported syntax was causing `gpbackup` to panic during the initial
table locking phase.

A TODO comment has been added to both modified functions (`LockTables`
and `LockTableNoWait`) to indicate that this is a workaround. The logic
should be reverted to use the `COORDINATOR ONLY` optimization once
Cloudberry implements this feature for feature and performance parity
with GPDB 7.

See https://github.com/apache/cloudberry/issues/1265