fix(test): Enable operator schema filter test on Cloudberry (#18)
The test for schema-filtered operators would previously fail when run
against a Cloudberry instance. This was because it relied on creating
a postfix (unary) operator, a feature not supported by Cloudberry.

This commit refactors the test to use a binary operator instead.
Binary operators are supported by both Greenplum and Cloudberry, which
allows the test for this core schema-filtering functionality to run
successfully on all platforms.

This change fixes a failing test and increases test coverage for
a core feature on Cloudberry.